+++ feature/articledetails/src/main/java/com/hshamkhani/articledetails/route/ArticleDetailsScreenRoute.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ArticleDetailsScreenRoute(
-    val articleId: Int,
+    val articleId: Long,
 )
